Fix text sizes

+++ src/main/java/org/pixle/client/render/entity/IEntityRenderer.java
@@ -4,5 +4,5 @@ import org.pixle.entity.Entity;
 import org.pixle.level.Level;
 
 public interface IEntityRenderer<T extends Entity> {
-    void render(T entity, int x, int y, Level level, float delta);
+    void render(T entity, float x, float y, Level level, float delta);
 }
